[feat] entity design (work in progress)
