+++ src/main/java/HDmedi/Server/ServerApplication.java
@@ -2,7 +2,9 @@ package HDmedi.Server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class ServerApplication {
