--- src/main/java/HDmedi/Server/domain/medicine_efficacy/repository/MedicineEfficacyRepository.java
@@ -1,7 +1,0 @@
-package HDmedi.Server.domain.medicine_efficacy.repository;
-
-import HDmedi.Server.domain.medicine_efficacy.entity.MedicineEfficacy;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MedicineEfficacyRepository extends JpaRepository<MedicineEfficacy, Long> {
-}
