--- src/main/java/HDmedi/Server/domain/medicine_warning/repository/MedicineWarningRepository.java
@@ -1,7 +1,0 @@
-package HDmedi.Server.domain.medicine_warning.repository;
-
-import HDmedi.Server.domain.medicine_warning.entity.MedicineWarning;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MedicineWarningRepository extends JpaRepository<MedicineWarning, Long> {
-}
