+++ src/main/java/HDmedi/Server/domain/user_entity/controller/HelloController.java
@@ -1,4 +1,4 @@
-package HDmedi.Server.domain.user.controller;
+package HDmedi.Server.domain.user_entity.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
