+++ src/main/java/HDmedi/Server/domain/user_entity/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package HDmedi.Server.domain.user.repository;
+package HDmedi.Server.domain.user_entity.repository;
 
-import HDmedi.Server.domain.user.entity.UserEntity;
+import HDmedi.Server.domain.user_entity.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
